+++ scripts/prep_examples.py
@@ -1,0 +1,6 @@
+def prep_examples():
+    pass
+
+
+if __name__ == "__main__":
+    prep_examples()
